% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_data.R
\docType{data}
\name{country_iso3}
\alias{country_iso3}
\title{Country abbreviations}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 181 rows and 7 columns.}
\usage{
country_iso3
}
\description{
ISO3 codes for countries produced from countrycode
}
\keyword{datasets}
