% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_covid_data.R
\name{prep_covid_raw_us}
\alias{prep_covid_raw_us}
\title{Download and prep US data}
\usage{
prep_covid_raw_us(cases_url, deaths_url)
}
\description{
Download and prep US data from JHU CSSE
}
\examples{
cases_url <- "https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_US.csv"
deaths_url <- "https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_deaths_US.csv"
cov_raw <- prep_covid_raw_us(cases_url,deaths_url)
}
