% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_covid_data.R
\name{prep_covid_raw_world}
\alias{prep_covid_raw_world}
\title{Download and prep world data}
\usage{
prep_covid_raw_world(cases_url, deaths_url)
}
\description{
Download and prep world data from JHU CSSE
}
\examples{
cov_raw <- prep_covid_raw_world()
}
