% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_data.R
\docType{data}
\name{rnatural_worldmap}
\alias{rnatural_worldmap}
\title{World map}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 177 rows and 2 columns.}
\usage{
rnatural_worldmap
}
\description{
ISO3 codes for countries produced from countrycode
}
\keyword{datasets}
