% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_data.R
\docType{data}
\name{us_states}
\alias{us_states}
\title{US States}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 52 rows and 4 columns.}
\usage{
us_states
}
\description{
Lat and lon info, abbreviations, for US states
}
\keyword{datasets}
