% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_data.R
\docType{data}
\name{usaboundaries_usstates}
\alias{usaboundaries_usstates}
\title{US state boundaries}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 52 rows and 12 columns.}
\usage{
usaboundaries_usstates
}
\description{
ISO3 codes for countries produced from countrycode
}
\keyword{datasets}
